+++ spec/PhpSpec/Listener/MethodReturnedNullListenerSpec.php
@@ -119,6 +119,8 @@ class MethodReturnedNullListenerSpec extends ObjectBehavior
         MethodCallEvent $methodCallEvent, ExampleEvent $exampleEvent, ConsoleIO $io, ResourceManager $resourceManager, SuiteEvent $event
     ) {
         $resourceManager->createResource(Argument::any())->willThrow(new \RuntimeException());
+        $methodCallEvent->getSubject()->willReturn(new \stdClass());
+        $methodCallEvent->getMethod()->willReturn('');
 
         $this->afterMethodCall($methodCallEvent);
         $this->afterExample($exampleEvent);
@@ -131,6 +133,8 @@ class MethodReturnedNullListenerSpec extends ObjectBehavior
         MethodCallEvent $methodCallEvent, ExampleEvent $exampleEvent, ConsoleIO $io, SuiteEvent $event
     ) {
         $io->isCodeGenerationEnabled()->willReturn(false);
+        $methodCallEvent->getSubject()->willReturn(new \stdClass());
+        $methodCallEvent->getMethod()->willReturn('');
 
         $this->afterMethodCall($methodCallEvent);
         $this->afterExample($exampleEvent);
@@ -167,6 +171,9 @@ class MethodReturnedNullListenerSpec extends ObjectBehavior
         $notEqualException->getExpected()->willReturn('foo');
         $notEqualException2->getExpected()->willReturn('bar');
 
+        $methodCallEvent->getSubject()->willReturn(new \stdClass());
+        $methodCallEvent->getMethod()->willReturn('');
+
         $this->afterMethodCall($methodCallEvent);
         $this->afterExample($exampleEvent);
 
@@ -182,6 +189,8 @@ class MethodReturnedNullListenerSpec extends ObjectBehavior
         MethodCallEvent $methodCallEvent, ExampleEvent $exampleEvent, ConsoleIO $io, SuiteEvent $event
     ) {
         $io->isFakingEnabled()->willReturn(false);
+        $methodCallEvent->getSubject()->willReturn(new \stdClass());
+        $methodCallEvent->getMethod()->willReturn('');
 
         $this->afterMethodCall($methodCallEvent);
         $this->afterExample($exampleEvent);
@@ -193,6 +202,9 @@ class MethodReturnedNullListenerSpec extends ObjectBehavior
     function it_prompts_when_correct_type_of_exception_is_thrown(
         MethodCallEvent $methodCallEvent, ExampleEvent $exampleEvent, ConsoleIO $io, SuiteEvent $event
     ) {
+        $methodCallEvent->getSubject()->willReturn(new \stdClass());
+        $methodCallEvent->getMethod()->willReturn('');
+
         $this->afterMethodCall($methodCallEvent);
         $this->afterExample($exampleEvent);
         $this->afterSuite($event);