+++ CHANGES.md
@@ -2,7 +2,7 @@ Upcoming 4.0
 ============
 
 * Dropped support for PHP versions less than 7.0 (@ciaranmcnulty)
-* Added scalar types and return types (@Sam-Burns)
+* Added scalar types and return types (@Sam-Burns, @ciaranmcnulty)
 
 3.4.0 / 2017-05-12
 ==================