+++ CHANGES.md
@@ -1,5 +1,5 @@
-Upcoming 2.x
-============
+2.5.8 / 2017-07-29
+==================
 
 * [fixed] parameters after extensions ignored in config file (@borNfreee)